style: fix checkstyle violations

+++ src/main/java/edu/sandiego/comp305/Group.java
@@ -5,13 +5,18 @@ import java.util.ArrayList;
 
 public class Group {
     private final String name;
+
     private final List<Team> teams;
+
     private final List<Standings> standings;
 
     public Group(final String name, final List<Team> teams){
         this.name = name;
+
         this.teams = new ArrayList<>(teams);
+
         this.standings = new ArrayList<>();
+
         for (final Team team : teams) {
             standings.add(new Standings(team));
         }
@@ -32,10 +37,10 @@ public class Group {
     public List<Standings> getStandings(){
         final List<Standings> sorted = new ArrayList<>(standings);
         sorted.sort((teamA, teamB) -> {
-            if (teamB.getPoints() != teamA.getPoints()) {
+            if (teamB.getPoints().intValue() != teamA.getPoints()) {
                 return teamB.getPoints() - teamA.getPoints();
             }
-            if (teamB.getGoalDiff() != teamA.getGoalDiff()) {
+            if (teamB.getGoalDiff().intValue() != teamA.getGoalDiff()) {
                 return teamB.getGoalDiff() - teamA.getGoalDiff();
             }
             return teamB.getGoalsFor() - teamA.getGoalsFor();
